chore: tweak structure a little bit
